refactor: move EmptyState to folder structure
Move EmptyState component from single file to empty-state/index.tsx
structure for better organization and consistency.
