+++ src/components/tasks/index.ts
@@ -1,4 +1,4 @@
 export { TasksList } from "./TasksList";
 export { TaskCard } from "./TaskCard";
-export { EmptyState } from "./EmptyState";
+export { EmptyState } from "./empty-state";
 export { LoadingState } from "./LoadingState";
